no more protected, pull out Integration
